Second simpler pass at interactions
