+++ lib/atomic_admin.rb
@@ -8,4 +8,6 @@ module AtomicAdmin
   mattr_accessor :admin_jwks_url
   mattr_accessor :audience
   mattr_accessor :internal_secret
+  mattr_accessor :application_interactions, default: AtomicAdmin::Interaction::Manager.new
+  mattr_accessor :application_instance_interactions, default: AtomicAdmin::Interaction::Manager.new
 end
