+++ lib/atomic_admin/schema.rb
@@ -1,17 +1,12 @@
-
 require_relative "schema/atomic_application_update_schema"
+require_relative "schema/application_instance_schema"
+require_relative "schema/application_instance_configuration_schema"
+require_relative "schema/application_instance_general_settings_schema"
+require_relative "schema/application_instance_xml_config_schema"
+require_relative "schema/application_instance_trial_details_schema"
+require_relative "schema/application_instance_license_details_schema"
 
 module AtomicAdmin
   module Schema
-    def self.for(resource, type)
-      case [resource.class.name, type]
-      when ["Application", "update"]
-        schema = AtomicApplicationUpdateSchema.for(resource)
-        # schema = Application.update_schema(schema)
-        schema
-      else
-        raise "Unknown resource type: #{resource.class.name}"
-      end
-    end
   end
 end
