+++ lib/atomic_admin/schema/application_instance_xml_config_schema.rb
@@ -1,0 +1,65 @@
+module AtomicAdmin::Schema
+  class ApplicationInstanceXmlConfigSchema < ApplicationInstanceSchema
+    def schema
+      {
+        type: "object",
+        properties: {
+          lti_key: {
+            type: "string",
+          },
+          lti_secret: {
+            type: "string",
+          },
+          lti_config_xml: {
+            type: "string",
+          },
+        }
+      }
+    end
+
+    def uischema
+      {
+        type: "Group",
+        label: "XML Configuration",
+        elements: [
+          {
+            type: "VerticalLayout",
+            elements: [
+              {
+                type: "Control",
+                scope: "#/properties/lti_key",
+                options: {
+                  props: {
+                    label: "LTI Key",
+                    isReadOnly: true,
+                  },
+                },
+              },
+              {
+                type: "Control",
+                scope: "#/properties/lti_secret",
+                options: {
+                  props: {
+                    label: "LTI Secret",
+                    isReadOnly: true,
+                  },
+                },
+              },
+              {
+                type: "Control",
+                scope: "#/properties/lti_config_xml",
+                options: {
+                  format: "textarea",
+                  props: {
+                    label: "LTI Config XML",
+                    isReadOnly: true,
+                  },
+                },
+              },
+            ],
+          },
+        ],
+      }
+    end
+  end
+end
